+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Fix: Run event processors and enrich transactions with contexts (#1430)
 * Bump: sentry-native to 0.4.9 (#1431)
 * Fix: Set user on transaction in Spring & Spring Boot integrations (#1443)
+* Ref: nullability annotations to Sentry module (#1439)
 
 # 4.4.0-alpha.2